+++ main.tf
@@ -31,6 +31,9 @@ module "vpc" {
   internal_ipv6_range                       = var.internal_ipv6_range
   network_firewall_policy_enforcement_order = var.network_firewall_policy_enforcement_order
   network_profile                           = var.network_profile
+  bgp_always_compare_med                    = var.bgp_always_compare_med
+  bgp_best_path_selection_mode              = var.bgp_best_path_selection_mode
+  bgp_inter_region_cost                     = var.bgp_inter_region_cost
 }
 
 /******************************************