+++ modules/fabric-net-svpc-access/versions.tf
@@ -29,9 +29,9 @@ terraform {
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-network:fabric-net-svpc-access/v10.0.0"
+    module_name = "blueprints/terraform/terraform-google-network:fabric-net-svpc-access/v11.1.1"
   }
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-network:fabric-net-svpc-access/v10.0.0"
+    module_name = "blueprints/terraform/terraform-google-network:fabric-net-svpc-access/v11.1.1"
   }
 }