+++ modules/subnets-beta/versions.tf
@@ -25,6 +25,6 @@ terraform {
   }
 
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-network:subnets-beta/v10.0.0"
+    module_name = "blueprints/terraform/terraform-google-network:subnets-beta/v11.1.1"
   }
 }