+++ modules/vpc/variables.tf
@@ -62,7 +62,7 @@ variable "mtu" {
 
 variable "enable_ipv6_ula" {
   type        = bool
-  description = "Enabled IPv6 ULA, this is a permenant change and cannot be undone! (default 'false')"
+  description = "Enabled IPv6 ULA, this is a permanent change and cannot be undone! (default 'false')"
   default     = false
 }
 
@@ -88,4 +88,22 @@ variable "network_profile" {
       * https://www.googleapis.com/compute/beta/projects/{projectId}/global/networkProfiles/{network_profile_name}
       * projects/{projectId}/global/networkProfiles/{network_profile_name}
     EOT
+}
+
+variable "bgp_always_compare_med" {
+  type        = bool
+  description = "If set to true, the Cloud Router will use MED values from the peer even if the AS paths differ. Default is false."
+  default     = false
+}
+
+variable "bgp_best_path_selection_mode" {
+  type        = string
+  description = "Specifies the BGP best path selection mode. Valid values are `STANDARD` or `LEGACY`. Default is `LEGACY`."
+  default     = "LEGACY"
+}
+
+variable "bgp_inter_region_cost" {
+  type        = string
+  description = "Specifies the BGP inter-region cost mode. Valid values are `DEFAULT` or `ADD_COST_TO_MED`."
+  default     = null
 }